Projectile JSON B
Fully implemented JSON based projectiles and converted the Ember projectile into JSON. The Ember Scepter is now broken but will be removed.
The next step is to implement projectiles for Equipment.
Fixed a bug where Yeti Meat could not be cooked.
Lowered the Underground spawner height to 50 and below from 60 and halved the spawn chance.
Renamed the "effect" creature json property to "effectDuration" and "amplifier" to "effectAmplifier".
The new Ember projectile has less range and only burns for 3 seconds.