Elements and Effects
Completed the Elements section of the Beastiary with descriptions of each Element as well as each beneficial and detrimental effect of that element.
Summoning Pedestals now have a much cheaper recipe but require Redstone Dust or Blocks as fuel.
Added Bleed effect which causes damage when moving.
Nymphs, Barghests and Gekens now all cause bleeding instead of their old effects.