Upgraded All Core Entity Classes
All core entity classes now compile, entity size has been moved to the new EntityType class so I will have to look into implementing my own and hooking it up with Creature Info and adapt the size scaling to it.
The Move Helpers, now called Movement Controllers are pretty much just renamed which is a relief as I have a lot of custom pathing code that I would not want to deal with on top of everything else!