Network Packets, Tile Entities, Containers and Some GUI
Updated the Network Packets.
Updated all Tile Entities.
Updated all Containers.
Updated most of the GUI, still some to go through.
Containers will need to be registered and I need to look at how they are opened now with the new Window ID stuff.
Overridden getSize() on EntityCreatureBase to dynamically add the size scale on top of the default EntityType. I will be removing the scaling options from the Creature JSON so that only the exact with and height can be set there for clarity.