Updated Mappings
Updated to the latest mappings as of this commit.
Improved the constructors of Creature entities so that they take an Entity Type, I will later do this to projectile classes as well. Sadly I'm stilling having this client side pig issue, from what I can see i think this is just beta forge and that I need to wait a bit.