Updated Equipment Forge
Figured out the issue with TileEntities not being created and fixed a bug where the wrong containers were being opened on the client.
The Equipment Forge is fully functional along with Equipment.
The Summoning Pedestal is still broken as I need to figure out how to use GUI Lists and Buttons, this is the last step towards fixing all GUI screen elements.
Fixed a bug where land creatures were unable to path when in water.