Biome Bleed Fix
Fixed a very old bug where mobs can spawn randomly in any biome so long as the biome they are restricted to is nearby. This wasn't too important in older versions of the mod but now in 1.14.3 the custom spawner is used for all spawning in this mod and needs it.
Fixed a bug where biome ids were being read too early by spawners meaning that modded biome ids were ignored (tags were fine).
Fixed a bug where the version checker and beastiary was unprepared for receiving no versions at all from the mod website.