AI Goal Prioritisation Fix
Fixed a bug where the order that AI goals were in was unpredictable causing mobs to behave incorrectly from inconsistent tempting to docile mobs. For some reason the default int values I used for the various AI index starts were being reset to 0, no idea how this is happening, but as a workaround they are now setup from the class constructor.