2.1.0.0 Beta 09
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.0-BETA09 for Minecraft 1.14.3
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
New Features:
--------------------
This is a BETA release of Lycanites Mobs for Minecraft 1.14.3.
The mod is very stable now but needs a bit more testing, hence Beta.


--------------------
Major Fixes:
--------------------
Fixed a crash when mounted on a creature with a debuff that the mount can cure.
Projectiles and special entities are now visible and fully functional.
Fixed the sky and storm spawners (Rocs, Zephyr, etc).


--------------------
Changes:
--------------------
Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
The Beastiary resets on death or dimension change.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
Behemoth and Pinky Models were removed and will be replaced with new models like the Belph soon.