2.1.0.0 Beta 15
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.0-BETA15 for Minecraft 1.14.4
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
New Features:
--------------------
This is a BETA release of Lycanites Mobs for Minecraft 1.14.4.
The mod is very stable now but needs a bit more testing, hence Beta.
Updated the mod to support Forge 28.0.45.
Re-implemented the summoning selection wheel (press and hold R by default).


--------------------
Major Fixes:
--------------------
Fixed a bug where mobs were spawning places that they don't fit causing a lot of weird things like underground Treants. This will need further testing to confirm the fix.


--------------------
Changes:
--------------------
Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon.


--------------------
Balancing:
--------------------
Reduced the leaping height of Wargs, Barghests and Maugs as well as their fall resistance.


--------------------
Minor Fixes:
--------------------
Fixed bugs when selecting a subspecies for summoning.
Fixed minion selection buttons not highlighting correctly when selected/hovered.