2.1.0.0 Beta 20
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.0-BETA20 for Minecraft 1.14.4
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
New Features:
--------------------
This is a BETA release of Lycanites Mobs for Minecraft 1.14.4.
The mod is very stable now but needs a bit more testing, hence Beta.

Updated to support recent Forge versions up to 28.0.100.


--------------------
Known Issues:
--------------------
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.


--------------------

Missing Features:
--------------------
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.