2.1.0.0 1.14.4 Initial Release
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.0 for Minecraft 1.14.4
============================================================

--------------------
New Features:
--------------------
I've decided that Lycanites Mobs is stable enough for a full release so it is no longer in beta for 1.14.4! That said, there are still some missing features and issues to work out.

 

--------------------
Changes:
--------------------
Increased minimum Forge requirement to 28.1.0, this is non-beta Forge so should be used over beta versions from now on.


--------------------

Known Issues:
--------------------
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.


--------------------

Missing Features (from 1.12.2):
--------------------
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.

 

--------------------
Balancing:
--------------------
Reduced the speed and health of Wraamon.
Created a seperate Fire spawner for the Nether (firenether.json), this spawns Cinders at a much slower rate in the Nether, the default fire spawner (fire.json) still runs at full speed in the overworld and the end as well as other dimensions. I found that there were too many Cinders in the Nether. Lava mobs still spawn at full speed as I feel these are balanced enough.
Reduced the spread range of the Plague debuff from 6 down to 2 blocks.

 

--------------------
Minor Fixes:
--------------------
Changing dimensions or teleporting long distances with a soulbound pet active no longer puts the pet into a respawn cooldown, instead the pet must actually die (health must hit 0) for a respawn timer to become active.
Fixed a couple of language errors in the Beastiary (garbage was displayed instead of empty strings as some of the page titles).