2.1.0.2 Subspecies Fixes
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.2 for Minecraft 1.14.4
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
Major Fixes:
--------------------
Fixed a hang (game/server lock up) caused by structure based spawners in the Nether or End, this is due to the structure locate bug (where using the /locate command in a dimension with no structures to locate will cause the game to hang.
Fixed a bug where rapid fire projectile attacks were incorrectly sending packets to clients.


--------------------
Config Changes:
--------------------
Added a config option to allow the despawning of rare subspecies.


--------------------
Changes:
--------------------


Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.


--------------------
Minor Fixes:
--------------------
Added an item drop lock so that creatures can only drop their death items once, this is a workaround for mob health conflicts that allow for unkillable mobs.
Rare subspecies elementals can now only fuse with other rare subspecies elementals.
Fixed possible crash when loading familiars.