2.1.0.4 Mob Event Dimension ID Fix
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.4 for Minecraft 1.14.4
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
Major Fixes:
--------------------
Fixed "Get Type Accessed Too Early!" log spam.
Fixed a crash with other mods when the player is invisible (Grue Claw Lv 3).


--------------------
Config Changes:
--------------------
Fixed default mob event jsons to use dimension ids (such as minecraft:overworld) instead of indexes (such as 0).
Added globalmobevent.json which allows for conditions to be added to every mobevent at once instead of individually in each mob event json file. This can be used to disable all events in a specific dimension, etc.


--------------------
Changes:
--------------------
Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.


--------------------
Minor Fixes:
--------------------
Fixed the Nether Portal spawner not working.