2.1.0.5 Xaphan Redstone Fix
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.5 for Minecraft 1.14.4
============================================================

--------------------
Major Fixes:
--------------------
Fixed a bug where the Assist pet stance was not saved to NBT.
Fixed a bug where redstone block updates could spawn Xaphans rapidly.

 

--------------------
Changes:
--------------------
Increased the spawn chance of Xaphans (the spawner is called less frequently due to the fix so they should spawn at about the same rate).


--------------------

Known Issues:
--------------------
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.


--------------------

Missing Features:
--------------------
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.

 

--------------------
Minor Fixes:
--------------------
Improved the accuracy of large ranged attacks such as Asmodeus' or Rahovart's direct attacks.
Improved compatibility of the Wisp and Sylph models as only their base model layer was rendering with certain graphical mods causing them to look really strange (a bald Wisp is a sad Wisp).