2.1.0.6 Water Mount Speed Fix
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.6 for Minecraft 1.14.4
============================================================

--------------------
Changes:
--------------------
Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.


--------------------
Minor Fixes:
--------------------
Water only mounts will now move at only 10% speed when on land.
Pets and minions set to Assist will now attack enemies hit by any attack from the player, previously they only responded to melee attacks.