2.1.0.7 Return of The Concapede!
============================================================
Lycanites Mobs Update: Reforged - Version 2.1.0.7 for Minecraft 1.14.4
Configs older than 2.0.0.0 will be reset.
============================================================

--------------------
Config Changes:
--------------------
Added "spawnDistanceMin" and "spawnDistanceMax" as new player spawn conditions for spawners to use, these allows for spawners to be restricted by how far the player is from the world spawn point.


--------------------
Changes:
--------------------
Concapedes now additionally spawn in Forest biomes by default.
Heavy mobs are now able to pickup other heavy mobs, this allows the Quetzodracl to pickup larger mobs like the Arisaur. Bosses can still never be picked up.
Entities with a width or height larger than 4 are now considered heavy for picking up.
Improved the proportions of the Belph and Behemoth models, thanks for feedback from Shivaxi's stream.

Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.
Entering the Nether can caused the game to hang as can starting a world with no structures, I've found that using the vanilla /locate command in worlds like this can also cause this so have determined this to be a vanilla/forge issue. A workaround for now is to disable the temple and village mob spawners.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.
The old Pinky Model was removed and will be replaced with new models like the Belph and Behemoth soon, for now Pinky spawning is disabled by default.


--------------------
Minor Fixes:
--------------------
Fixed a bug where dungeon bosses were tameable.
Fixed Alpha Jousts not being hostile to players and villagers.
When a soulbound pet's owner dies, the pet will now save itself to nbt before despawning to ensure that their inventory is updated rather than lost.
Quetzodracl and Morocks are now flagged as heavy (cannot be picked up by Rocs, etc).
Set the Paddle AI Goal of water/lava breathing mobs to allow sinking, this should stop some water mobs from just bobbing on the surface.
Fixed Cinders not spawning in the nether due to the new netherfire.json spawner not generating correctly.