2.1.2.0 IMProvements ============================================================ Lycanites Mobs Update: IMProvements! - Version 2.1.2.0 for Minecraft 1.14.4 Configs older than 2.0.0.0 will be reset. ============================================================
--------------------
Major Fixes:
--------------------
New and IMProved Afrit and Arix Imp models!


--------------------
Config Changes:
--------------------
Fixed a typo with griefing creature json flags.


--------------------
Changes:
--------------------
The Wildkin has been moved to the Brute Type and must now be tamed with Brute Treats.

Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.