2.0.4.3 New Zoataur ============================================================ Lycanites Mobs Update: Avian Awesomeness - Version 2.0.4.3 for Minecraft 1.12.2 Configs older than 2.0.0.0 will be reset. ============================================================
--------------------
New Features:
--------------------
Added a brand new redesigned Zoataur! It is now a Dragon that can spawn in Plains and Deserts with a reduced spawn chance and increased health. It can now also be tamed as a mount with a block ability that will protect the rider!


--------------------
Major Fixes:
--------------------
Fixed a random crash caused by large mobs melee attacking.