2.1.3.5 New Zoataur ============================================================ Lycanites Mobs Update: Avian Awesomeness - Version 2.1.3.5 for Minecraft 1.14.4 Configs older than 2.0.0.0 will be reset. ============================================================
--------------------
New Features:
--------------------
Added a brand new redesigned Zoataur! It is now a Dragon that can spawn in Plains and Deserts with a reduced spawn chance and increased health. It can now also be tamed as a mount with a block ability that will protect the rider!


--------------------
Major Fixes:
--------------------
Fixed a random crash caused by large mobs melee attacking.


--------------------
Changes:
--------------------
Known Issues:
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.

Missing Features:
Fluids (Ooze, Pure Lava, etc) are not added back yet.