2.1.4.3 Projectile Fix ============================================================ Lycanites Mobs Update: RUN - Version 2.1.4.3 for Minecraft 1.14.4 ============================================================
--------------------
Major Fixes:
--------------------
Fixed projects instantly vanishing on the client side.
Fixed a bug where Equipment and Projectiles with summoning effects were not actually summoning anything.

 

--------------------
Changes:
--------------------
Improved the aiming of the Summoning Staff Portal and laser projectiles.


--------------------

Known Issues:
--------------------
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.


--------------------

Missing Features:
--------------------
Fluids (Ooze, Pure Lava, etc) are not added back yet.