2.2.0.0 First 1.15.1 Release ============================================================ Lycanites Mobs Update: RUN - Version 2.2.0.0 for Minecraft 1.15.1 Configs older than 2.0.0.0 will be reset. ============================================================
--------------------
New Features:
--------------------
First release of Lycanites Mobs for Minecraft 1.15.1!
There are some known rendering issues which I am working on but the mod should be as stable as 1.14.4 versions in other areas.
New laser rendering.

 

--------------------
Known Issues:
--------------------
Additive, Subtractive and Shadeless mob rendering is not re-implemented yet, this will cause some models such as Wisps or Zephyr to not look as good.

Glowing and stealth effects are not fully implemented yet.
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.


--------------------

Missing Features:
--------------------
Fluids (Ooze, Pure Lava, etc) are not added back yet.