============================================================ Lycanites Mobs Update: Fully Charged! - Version 2.2.1.4 for Minecraft 1.15.1 ============================================================
--------------------
Major Fixes:
--------------------
Fixed creature level only affect their Health and not their other stats.


--------------------
Known Issues:
--------------------
Effect blocks like Scorchfire or Frostwebs have a solid black background instead of an alpha cutout.
Dungeon Generation is fractured, the Dungeons do generate but with missing chunks as other WorldGen cuts through it, this is because they are generating as Features instead of Structures, this is temporary and will be fixed one Structure WorldGen is working.


--------------------
Missing Features:
--------------------
Fluids (Ooze, Pure Lava, etc) are not generated yet but seem to be working now.


--------------------
Minor Fixes:
--------------------
Fixed the Beastiary Pet tab not correctly displaying active pet health or experience.
Items stored in the Equipment Forge or Equipment Infuser are not saved or dropped if the block is destroyed.
Fixed a bug where firing a projectile at an entity within melee range was not working.
Fixed freshly tamed or summoned creatures having 3X health.